refactor
Split everything into separate files to try and make it more extensible and readable.
