+++ src/models/index.ts
@@ -1,0 +1,2 @@
+export { Tag } from './tag';
+export { Tags } from './tags';
