+++ src/services/index.ts
@@ -1,0 +1,5 @@
+export { loadDictionaries } from "./dictionary-services";
+export { getDictionaryPaths } from "./dictionary-services";
+export { loadConfiguredOrDefaultDictionaries } from "./dictionary-services";
+
+export { watchDictionaryFiles } from "./watcher-services";
