+++ src/utils/format-utils.ts
@@ -1,0 +1,4 @@
+
+export function canonicalLowerCase(str: string): string {
+    return str.normalize('NFD').toLocaleLowerCase().normalize('NFD');
+}
