+++ src/utils/index.ts
@@ -1,0 +1,4 @@
+export { stringToLineLengths } from './file-utils';
+export { lineNumberFromIndex } from './file-utils';
+
+export { canonicalLowerCase } from './format-utils';
